requirements.txt
pandas==2.3.3
openai
tqdm
numpy
python-dotenv